Refctored type generator singular property mapping
