+++ src/generators/types.ts
@@ -6,13 +6,13 @@ import { convertToPascalCase } from '@/src/utils/slug';
 
 import type {
   InterfaceDeclaration,
-  PropertySignature,
   TypeAliasDeclaration,
   TypeNode,
   TypeParameterDeclaration,
 } from 'typescript';
 
 import type { Model, ModelField } from '@/src/types/model';
+import { mapRoninFieldToTypeNode } from '@/src/utils/types';
 
 const DEFAULT_FIELD_SLUGS = [
   'id',
@@ -114,84 +114,14 @@ export const generateTypes = (
         factory.createTypeLiteralNode(
           fields
             .sort((a, b) => a.slug.localeCompare(b.slug))
-            .map((field) => {
-              const propertyUnionTypes = new Array<TypeNode>();
-
-              switch (field.type) {
-                case 'link': {
-                  // Check to make sure the target model exists. If it doesn't we
-                  // fall back to using `unknown` as the type.
-                  const targetModel = models.find((model) => model.slug === field.target);
-                  if (!targetModel) {
-                    propertyUnionTypes.push(
-                      factory.createKeywordTypeNode(SyntaxKind.UnknownKeyword),
-                    );
-                    break;
-                  }
-
-                  // If the field is marked as `many` then we need to wrap the
-                  // type in an array.
-                  const schemaTypeRef = factory.createTypeReferenceNode(
-                    convertToPascalCase(targetModel.slug),
-                  );
-                  const resolvedLinkFieldNode = factory.createTypeReferenceNode(
-                    identifiers.utils.resolveSchema,
-                    [
-                      field.kind === 'many'
-                        ? factory.createTypeReferenceNode(identifiers.primitive.array, [
-                            schemaTypeRef,
-                          ])
-                        : schemaTypeRef,
-
-                      factory.createTypeReferenceNode(genericIdentifiers.using),
-
-                      factory.createLiteralTypeNode(
-                        factory.createStringLiteral(field.slug),
-                      ),
-                    ],
-                  );
-
-                  propertyUnionTypes.push(resolvedLinkFieldNode);
-                  break;
-                }
-                case 'blob':
-                case 'boolean':
-                case 'date':
-                case 'json':
-                case 'number':
-                case 'string': {
-                  const primitive = MODEL_TYPE_TO_SYNTAX_KIND_KEYWORD[field.type];
-                  propertyUnionTypes.push(primitive);
-                  break;
-                }
-                default: {
-                  propertyUnionTypes.push(
-                    factory.createKeywordTypeNode(SyntaxKind.UnknownKeyword),
-                  );
-                  break;
-                }
-              }
-
-              // We need to mark fields as nullable if they are:
-              // - Not required
-              // - Not a link field
-              // - Not a many-to-many link field
-              if (
-                field.required === false &&
-                field.type === 'link' &&
-                field.kind !== 'many'
-              )
-                propertyUnionTypes.push(
-                  factory.createLiteralTypeNode(factory.createNull()),
-                );
-
-              return factory.createPropertySignature(
+            .map((field) =>
+              factory.createPropertySignature(
                 undefined,
                 field.slug,
                 undefined,
-                factory.createUnionTypeNode(propertyUnionTypes),
-              );
-            })
+                factory.createUnionTypeNode(mapRoninFieldToTypeNode(field, models)),
+              ),
+            )
             .filter(Boolean),
         ),
       ]),
